SDR Moved from ntypes to types
renamed: src/nupic/ntypes/Sdr.hpp        -> src/nupic/types/Sdr.hpp
renamed: src/nupic/ntypes/SdrProxy.hpp   -> src/nupic/types/SdrProxy.hpp
renamed: src/nupic/ntypes/SdrMetrics.hpp -> src/nupic/utils/SdrMetrics.hpp
