+++ src/nupic/algorithms/Connections.hpp
@@ -34,7 +34,7 @@
 #include <nupic/math/Math.hpp>
 #include <nupic/types/Types.hpp>
 #include <nupic/types/Serializable.hpp>
-#include <nupic/ntypes/Sdr.hpp>
+#include <nupic/types/Sdr.hpp>
 
 namespace nupic {
