+++ src/nupic/types/Sdr.hpp
