+++ src/nupic/types/SdrProxy.hpp
@@ -23,7 +23,7 @@
 #define SDR_PROXY_HPP
 
 #include <vector>
-#include <nupic/ntypes/Sdr.hpp>
+#include <nupic/types/Sdr.hpp>
 #include <nupic/types/Types.hpp>
 #include <nupic/types/Serializable.hpp>
