+++ src/nupic/utils/SdrMetrics.hpp
@@ -24,7 +24,7 @@
 
 #include <vector>
 #include <numeric>
-#include <nupic/ntypes/Sdr.hpp>
+#include <nupic/types/Sdr.hpp>
 #include <nupic/types/Types.hpp>
 #include <nupic/types/Serializable.hpp>
 #include <nupic/utils/Random.hpp>
