+++ src/test/unit/ntypes/SdrProxyTest.cpp
@@ -17,8 +17,8 @@
 
 #include <vector>
 #include <gtest/gtest.h>
-#include <nupic/ntypes/Sdr.hpp>
-#include <nupic/ntypes/SdrProxy.hpp>
+#include <nupic/types/Sdr.hpp>
+#include <nupic/types/SdrProxy.hpp>
 
 using namespace std;
 using namespace nupic;
